Add dbtemplate package
This is modeled after the docker package to make it easy to use as a
replacement. Instead of creating a new docker container for each test,
it will create a new database in an existing database container, using a
template database as a source. This assumes that the template database
exists and already has the schema migrations run against it.

This should probably be refactored more so the inputs and options are
relevant, instead of following the pattern from the docker package. But
it works as an inital proof of concept.