Refactor tests to use dbtemplate package instead of docker
This swaps the test initialization to use the dbtemplate package so each
test will get a fresh database created from a template database, instead
of a completely new docker container.

This also disables (via commenting out) or adds skips to some tests that
do not pass after this refactor. The tests fail due to the initial state
of a test not being an empty database. Now the inital state is a
database with all migrations already run. These tests would need to be
refactored more, or the dbtemplate package refactored to allow creation
from a template without migrations.

However, some of these test might also be better suited for the pure sql
tests that are being implemented.