Update additional tests to use dbtemplate
This changes how the api tests setup works to use the dbtemplate
approach instead of docker. However, this particular change probably is
not correct since the change happens in a place that would impact
`boundary dev`. Some additional refactoring needs to be done to allow
tests to use dbtemplate without impacting non-test code.